+++ Collections/Presentation/Presentation-Domain/Screen/Home/HomeViewController.swift
@@ -5,7 +5,8 @@ final class HomeViewController: UIViewController {
     @IBOutlet weak var tableView: UITableView!
 
     private let router: RouterProtocol = Router()
-    private let homeItems = CellItem.allCases.compactMap { $0.rawValue }
+
+    private var dataSource: HomeDataSource! = HomeDataSource()
 
     static func createInstance() -> HomeViewController {
         HomeViewController.instantiateInitialViewController()
@@ -21,31 +22,7 @@ extension HomeViewController {
 
     private func setupTableView() {
         tableView.register(HomeTableViewCell.xib(), forCellReuseIdentifier: HomeTableViewCell.resourceName)
-        tableView.delegate = self
-        tableView.dataSource = self
-    }
-}
-
-extension HomeViewController: UITableViewDelegate {
-
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 100
-    }
-}
-
-extension HomeViewController: UITableViewDataSource {
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return homeItems.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(
-                withIdentifier: HomeTableViewCell.resourceName,
-                for: indexPath
-        ) as? HomeTableViewCell ?? HomeTableViewCell()
-
-        cell.setup(item: homeItems[indexPath.row])
-        return cell
+        tableView.dataSource = dataSource
+        tableView.rowHeight = 100
     }
 }
